+++ go-fiber-mysql/internal/app/responses/user_response.go
@@ -7,7 +7,7 @@ import (
 )
 
 type UserResponse struct {
-	ID          uint       `json:"id"`
+	ID          uint64     `json:"id"`
 	Name        string     `json:"name,omitempty"`
 	Email       string     `json:"email,omitempty"`
 	LastLoginAt *time.Time `json:"last_login_at,omitempty"`
@@ -15,11 +15,11 @@ type UserResponse struct {
 	Roles []RoleResponse `json:"roles,omitempty"`
 
 	CreatedAt   *string `json:"created_at,omitempty"`
-	CreatedByID *uint   `json:"created_by_id,omitempty"`
+	CreatedByID *uint64 `json:"created_by_id,omitempty"`
 	UpdatedAt   *string `json:"updated_at,omitempty"`
-	UpdatedByID *uint   `json:"updated_by_id,omitempty"`
+	UpdatedByID *uint64 `json:"updated_by_id,omitempty"`
 	DeletedAt   *string `json:"deleted_at,omitempty"`
-	DeletedByID *uint   `json:"deleted_by_id,omitempty"`
+	DeletedByID *uint64 `json:"deleted_by_id,omitempty"`
 }
 
 func NewUserResponse(data *entities.User) UserResponse {
